Cleaned up Java8 problem 4
Cleaned up Java 8, problem 4

+++ java8/src/util/MathUtils.java
@@ -7,7 +7,7 @@ import java.util.stream.LongStream;
  * Created by gcharles on 2/17/17.
  */
 public class MathUtils {
-    public static boolean isPalindrome(int n) {
+    public static boolean isPalindrome(long n) {
         String s = String.valueOf(n);
         return new StringBuilder(s).reverse().toString().equals(s);
     }
